+++ pkg/piecestore/rpc/client/grpc.go
@@ -1,7 +1,7 @@
 // Copyright (C) 2018 Storj Labs, Inc.
 // See LICENSE for copying information.
 
-package ranger
+package client
 
 import (
 	"bytes"
@@ -9,18 +9,22 @@ import (
 	"io"
 	"io/ioutil"
 
-	"storj.io/storj/pkg/piecestore/rpc/client"
+	"github.com/zeebo/errs"
+
+	"storj.io/storj/pkg/ranger"
 )
 
+var Error = errs.Class("grpcRanger error")
+
 type grpcRanger struct {
-	c    *client.Client
+	c    *Client
 	id   string
 	size int64
 }
 
 // GRPCRanger turns a gRPC connection to piece store into a Ranger
-func GRPCRanger(ctx context.Context, c *client.Client, id string) (Ranger, error) {
-	piece, err := c.Meta(ctx, client.PieceID(id))
+func GRPCRanger(ctx context.Context, c *Client, id string) (ranger.Ranger, error) {
+	piece, err := c.Meta(ctx, PieceID(id))
 	if err != nil {
 		return nil, err
 	}
@@ -30,7 +34,7 @@ func GRPCRanger(ctx context.Context, c *client.Client, id string) (Ranger, error
 // GRPCRangerSize creates a GRPCRanger with known size.
 // Use it if you know the piece size. This will safe the extra request for
 // retrieving the piece size from the piece storage.
-func GRPCRangerSize(c *client.Client, id string, size int64) Ranger {
+func GRPCRangerSize(c *Client, id string, size int64) ranger.Ranger {
 	return &grpcRanger{c: c, id: id, size: size}
 }
 
@@ -53,7 +57,7 @@ func (r *grpcRanger) Range(ctx context.Context, offset, length int64) (io.ReadCl
 	if length == 0 {
 		return ioutil.NopCloser(bytes.NewReader([]byte{})), nil
 	}
-	reader, err := r.c.Get(ctx, client.PieceID(r.id), offset, length)
+	reader, err := r.c.Get(ctx, PieceID(r.id), offset, length)
 	if err != nil {
 		return nil, Error.Wrap(err)
 	}
